% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_nuts.R
\name{aggregate_to_nuts}
\alias{aggregate_to_nuts}
\title{Aggregate pre-processed data to the desired NUTS level}
\usage{
aggregate_to_nuts(level, ref, dat)
}
\arguments{
\item{level}{The desired NUTS level to aggregate data to. Given as \code{numeric}.}

\item{ref}{The reference table for joining. This is the output of \code{\link[suspendr:nuts_table]{nuts_table()}}.}

\item{dat}{Object of class \code{data.frame}. The data set with country specific case data. Needs to include columns with the following
names:
\describe{
\item{date}{A vector of class \code{Date} with the given dates.}
\item{new_cases}{A vector of class \code{numeric}. Contains the new cases of a given day and region.}
\item{lvl_3}{A vector of class \code{character}. Contains the NUTS 3 code of the region.}
\item{lvl_2}{A vector of class \code{character}. Contains the NUTS 2 code of the region.}
\item{lvl_1}{A vector of class \code{character}. Contains the NUTS 1 code of the region.}
\item{lvl_0}{A vector of class \code{character}. Contains the NUTS 0 code of the region.}
}
The order does not matter.}
}
\value{
A \code{data.frame} with the country specific case data aggregated to the desired
NUTS level.
}
\description{
Aggregate pre-processed data to the desired NUTS level
}
