% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_austria_cases.R
\name{get_austria_cases}
\alias{get_austria_cases}
\title{Get Austrian case data from AGES website or from cached file.}
\usage{
get_austria_cases(level, ref, cache_dir = NULL)
}
\arguments{
\item{level}{The desired NUTS-level. Given as object of class \code{numeric}.
One of \code{c(0, 1, 2, 3)}.}

\item{ref}{The reference table. The output of \code{\link[suspendr:nuts_table]{nuts_table()}}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{data.frame} with the country specific case data aggregated to the desired
NUTS level.
}
\description{
Get Austrian case data from AGES website or from cached file.
}
