% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_austria_cases.R
\name{get_austria_cases_from_source}
\alias{get_austria_cases_from_source}
\title{Get pre-processed Austrian COVID-19 case data from the AGES dashboard website.}
\usage{
get_austria_cases_from_source(cache_dir, filename)
}
\arguments{
\item{cache_dir}{The directory where to save the output of the function.}

\item{filename}{What the name of the file to cache is.}
}
\value{
A \code{data.frame} containing Austrian case data.
}
\description{
This function gets the Austrian case data from AGES. And returns
it aggregated to the desired NUTS level.
}
