% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CH_FL_cases.R
\name{get_CH_FL_cases}
\alias{get_CH_FL_cases}
\title{Get Swiss and Liechtensteinian COVID-19 case data from source or from cache.}
\usage{
get_CH_FL_cases(level, merge_li, ref, cache_dir = NULL)
}
\arguments{
\item{level}{A vector of class \code{numeric} and length 1. Indicates the desired NUTS level.}

\item{merge_li}{A binary indicator whether to merge Liechtenstein to the canton of St. Gallen on NUTS
level 3. This also implicates adding the population as well as cases to whichever region St. Gallen is in.}

\item{ref}{The reference table. The output of \code{\link[suspendr:nuts_table]{nuts_table()}}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{data.frame} with a time series of COVID-19 cases in Switzerland/Liechtenstein
on the desired NUTS level.
}
\description{
Get Swiss and Liechtensteinian COVID-19 case data from source or from cache.
}
