% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IOM_data.R
\name{get_country_lookup}
\alias{get_country_lookup}
\title{Get ISO-3 country codes used in the IOM travel restriction matrix}
\usage{
get_country_lookup(path)
}
\arguments{
\item{path}{The path to one of the pages displaying a concrete IOM travel restriction matrix.}
}
\value{
A \code{data.frame} with ISO-3 country codes and the countries' full name.
}
\description{
Get ISO-3 country codes used in the IOM travel restriction matrix
}
