% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germany_cases.R
\name{get_germany_cases_from_source}
\alias{get_germany_cases_from_source}
\title{Get German case data from RKI and pre-process it}
\usage{
get_germany_cases_from_source(ref, cache_dir, filename)
}
\arguments{
\item{ref}{The reference table. The output of \code{\link[suspendr:nuts_table]{nuts_table()}}.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}

\item{filename}{The name of the pre-processed data file that is save to \code{cache_dir}.}
}
\value{
A \code{data.frame} with pre-processed data originally from RKI
}
\description{
Get German case data from RKI and pre-process it
}
