% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IOM_data.R
\name{get_iom_dates}
\alias{get_iom_dates}
\title{Get the dates for which IOM travel restriction matrices are available}
\usage{
get_iom_dates(path)
}
\arguments{
\item{path}{The path to the overview page which shows for what dates IOM travel matrices are available.}
}
\value{
A vector of class \code{Date} containing dates for which IOM travel matrices are available.
}
\description{
Get the dates for which IOM travel restriction matrices are available
}
