% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IOM_data.R
\name{get_iom_matrix}
\alias{get_iom_matrix}
\title{Get IOM travel restriction matrix data}
\usage{
get_iom_matrix(countries = NULL, date, cache_dir = NULL)
}
\arguments{
\item{countries}{A vector of class \code{character}. Valid inputs are ISO-3 country codes (case insensitive).
Defaults to all countries.}

\item{date}{A vector of class \code{Date}. Matrices are returned for each day in \code{dates} on which
such a matrix is available.}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory")}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{list} containing processed IOM travel restriction matrices for days requested
in \code{date} if available. Rows refer to countries imposing restrictions to travellers/citizens from
countries in the columns.
}
\description{
Get IOM travel restriction matrix data
}
