% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IOM_data.R
\name{get_iom_matrix_color_code}
\alias{get_iom_matrix_color_code}
\title{Convert raw IOM travel matrix to matrix of color codes.}
\usage{
get_iom_matrix_color_code(mat_raw)
}
\arguments{
\item{mat_raw}{The output of \code{get_iom_matrix_raw}.}
}
\value{
A \code{data.frame} with recoded color codes of the IOM travel restriction matrix.
}
\description{
Convert raw IOM travel matrix to matrix of color codes.
}
