% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IOM_data.R
\name{get_iom_matrix_raw}
\alias{get_iom_matrix_raw}
\title{Get raw IOM travel restriction matrix}
\usage{
get_iom_matrix_raw(path)
}
\arguments{
\item{path}{The path to the javascript file where the matrix is defined.}
}
\value{
A data frame containing the level of restriction as well as the codes for concrete measures imposed.
}
\description{
Get raw IOM travel restriction matrix
}
