% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\name{pop_data}
\alias{pop_data}
\title{Get population data from Eurostat}
\usage{
pop_data(
  country = c("at", "austria", "de", "germany", "fr", "france", "it", "italy", "ch",
    "switzerland", "li", "liechtenstein", "all"),
  nuts = 2,
  cache_dir = NULL
)
}
\arguments{
\item{country}{Object of class \code{character}. One or more of
\code{c("at", "austria", "de", "germany", "fr", "france", "it", "italy", "ch",
"switzerland", "li", "liechtenstein", "all")}.}

\item{nuts}{Object of class \code{numeric} and length 1. Indicates the NUTS level}

\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory"}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
A \code{data.frame} containing three columns:
\describe{
\item{nuts_id}{The NUTS IDs of the regions in the specified countries.}
\item{time}{The most recent year which data is available for. Reference date is
the 31st of December.}
\item{population}{The number of inhabitants of the respective NUTS region.}
}
}
\description{
This function imports population data from Eurostat.
}
\examples{
pop_data(country = c("at", "ch"), nuts = 2)

}
