% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\name{process_raw_pop_data}
\alias{process_raw_pop_data}
\title{Process raw population data}
\usage{
process_raw_pop_data(raw_data, nuts, regex)
}
\arguments{
\item{raw_data}{The preprocessed data file. Basically this is the output from
\code{\link[suspendr:get_pop_data_from_source]{get_pop_data_from_source()}}, whether it comes from a cached file or from
the function itself.}

\item{nuts}{The nuts level specified in the main function.}

\item{regex}{The regular expressions corresponding to the countries specified in the main
function.}
}
\value{
A \code{data.frame} containing three columns:
\describe{
\item{nuts_id}{The NUTS IDs of the regions in the specified countries.}
\item{time}{The most recent year which data is available for. Reference date is
the 31st of December.}
\item{population}{The number of inhabitants of the respective NUTS region.}
}
}
\description{
Process raw population data
}
