% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts_table.R
\name{get_nuts_table_from_source}
\alias{get_nuts_table_from_source}
\title{Data set containing NUTS-Classifications}
\usage{
get_nuts_table_from_source(cache_dir = tempdir())
}
\arguments{
\item{cache_dir}{Object of class \code{character}. Indicates a path where the return object is cached. Defaults to \code{\link{tempdir}}.}
}
\value{
An object of class \code{data.frame} containing NUTS-Codes and names of the regions considered in the SUSPend project.
}
\description{
A function that returns a reference table for NUTS regions including IDs and names.
}
