% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\name{get_pop_data_from_source}
\alias{get_pop_data_from_source}
\title{Get population data from Eurostat}
\usage{
get_pop_data_from_source(path_to_file, regex_all)
}
\arguments{
\item{path_to_file}{Character string indicating where to save the pre-processed file.}

\item{regex_all}{Character string. Contains a regular expression to filter out countries
we are not interested in from the original data file from Eurostat.}
}
\value{
Returns a \code{data.frame} containing
}
\description{
Get population data from Eurostat
}
