% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts_table.R
\name{nuts_table}
\alias{nuts_table}
\title{Data set containing NUTS-Classifications}
\usage{
nuts_table(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{This argument indicates the path to the
desired cache directory. Defaults to \code{NULL}.
The desired cache directory is determined according to the following hierarchy:
\enumerate{
\item{If \code{cache_dir} defined by user: Caches files to user defined directory.}
\item{If \code{cache_dir == NULL} (default): check if \code{getOption("suspendr_cache_dir")} is
defined and use this as cache directory.}
\item{If \code{getOption("suspendr_cache_dir")} undefined: Use \code{tempdir()} as
cache directory.}
}
If you would like to always use the same cache directory, the simplest way to do so is
to run the following line of code: \code{options(suspendr_cache_dir = "path/to/your/cache/directory"}.
Afterwards you can just ignore the \code{cache_dir} argument as it always will use the path you set in
the options.}
}
\value{
An object of class \code{data.frame} containing NUTS-Codes and names of the regions considered in the SUSPend project.
}
\description{
A function that returns a reference table of NUTS regions including
IDs and names. The data is cached in \code{cache_dir} and read from there if the file
already exists. If the cached file is older than a week it is updated from the data sources.
}
\examples{
\donttest{nuts_table()}
}
\seealso{
\code{\link[base:tempdir]{tempdir()}}
}
